HTML, CSS, PHP, JS
This project is a representation of a platform to sell art, which is an E-commerce Website. The project contains some functionalities including crud operations like a normal E-commerce website can do. For example, here such as buyer can buy art, Artist can sell art, including a delivery guy who can deliver the arts to the buyers within a given time and an Admin to run the website operation. And its main mission is to support the art community of Bangladesh so that the artists can get their recognition in society and can earn their livelihood. Moreover, the upcoming generation can choose Art as a profession rather than a hobby so that the Art cultural of this country get secured.